planus-ai-app/requirements.txt
streamlit
pandas
google-generativeai
weasyprint
kaleido==0.1.0post1
plotly
pypdf
